ci(actions): implement caching for actions
Where possible, caching is used with nox to reduce action run-time. Here, it is implemented for linting, formatting, and security checks.

It is difficult tod ecide what the best way to implement this for testing is. Since I use nox to spin off eavh python version, and then poetry to install the project, any existing virtualenvs get clobbered by poetry on the install.