fix(repo): support python3.10
As Python 3.10.1 has released, and as pandas and numpy are now 3.10 compatible, we add support for python3.10.

Pre-commit and ReadTheDocs now uses py3.10 as default. Nox includes py3.10 in its matrix, and defaults to py3.10 for steps only run in one version. Pyproject.toml now explicitly states python versions. Also, though is technically a build release, I'm tagging as a fix to trigger the initial public release to pypi.