feat(cli): generalise data reading
By allowing the user to specify the number of rows in the metadata and the name of each field, it is possible to greatly increase the flexibility of the data read in. Additionally, this structure should ease future maintenance should the expected data input change or should more fields be required.

BREAKING CHANGE: This introduces 2 new command line flags and changes how the data is processed. Past scripts are not guaranteed to run correctly - unless, by chance, they used the defaults - so a breaking release is required.