fix(pipeline.py): correct enfc implementation
ENFC for lipid classes is now calculated on the summed total of all lipids in that class.

BREAKING CHANGE: Prior implementations had calculated the ENFC for class by taking the mean and st.dev of the individual lipid ENFC for all lipids in a class. While not incorrect, this was inconsistent with prior work of this framework. Closes #23.