build(pyproject.toml): pin numpy to ^1.22.2
Security update revealed a number of numpy security issues. To update to 1.22.2 (which has patches for all known, we must remove the python 3.7 compatibility. Given that the last bugfix for 3.7 was in 2020, this seems a fair trade to keep up to date on the analysis software. Additionally, we must pass an ignore to Safety, as their monthly database hasn't updated to show that 1.22.2 is patched.

See https://github.com/pyupio/safety/issues/364 and https://github.com/numpy/numpy/issues/19038